finished delete method
